requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
enum34==1.1.10
future==0.18.2
importlib-metadata==3.3.0
iniconfig==1.1.1
lxml==4.6.2
more-itertools==8.6.0
namedlist==1.8
numpy==1.19.4
packaging==20.8
pandas==1.1.5
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.4
selenium==3.141.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
waiting==1.4.1
wcwidth==0.2.5
webium==1.2.1
xlrd==2.0.1
zipp==3.4.0
selenium-wire==3.0.4
